generate ref docs
