+++ .gitignore
@@ -13,4 +13,5 @@ target/
 
 # ignore vim backup files
 *.sw[op]
+.DS_Store
