+++ website/docs/reference/cli-options.md
@@ -1404,6 +1404,10 @@ Available in commands:
 
 <!-- Automatically generated, DO NOT EDIT MANUALLY -->
 
+### `--args-file`
+
+File with scalac options.
+
 ### `--scalac-option`
 
 Aliases: `-O`, `--scala-opt`, `--scala-option`
